broker-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "broker-service"
version = "0.0.1"
description = "FastAPI broker microservice for cTrader Open API"
authors = [{ name = "Sebastian Stolz" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.28.0",
  "pydantic-settings>=2.2.1",
  "redis>=5.0.1",
  "ctrader-open-api>=0.0.28",
  "tenacity>=8.2.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "httpx>=0.25",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "pytest-mock>=3.12",
  "pytest-timeout>=2.2",
  "fakeredis[asyncio]>=2.23",
  "freezegun>=1.4"
]

[project.scripts]
broker-service = "app.main:run"
